remove files that are no longer used
