+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules
-app/credentials/*
+credentials/*
 serviceAccountKey.json
